Fixed database Get function by setting notifications to be local

+++ api/notification/notificationInput.go
@@ -2,7 +2,7 @@ package notification
 
 type NotificationInput struct {
 	URL     string `json:"url"`
-	Timeout int    `json:"timeout"`
+	Timeout int64  `json:"timeout"`
 	Field   string `json:"field"`
 	Country string `json:"country"`
 	Trigger string `json:"trigger"`
